added claim sub routing and refactored claims table
