+++ client/src/Dashboard/pages/Claims/components/ClaimHeader/ClaimHeader.js
@@ -6,6 +6,7 @@ import {
 	Button,
 	Divider
 } from "@material-ui/core";
+import { Link } from "react-router-dom";
 
 import styles from "./header.style.js";
 const ClaimHeader = (props) => {
@@ -19,7 +20,12 @@ const ClaimHeader = (props) => {
 					</Typography>
 				</Grid>
 				<Grid item xs={12} sm={2} className={classes.buttonContainer}>
-					<Button variant="contained" className={classes.button}>
+					<Button
+						variant="contained"
+						className={classes.button}
+						to="/admin/claims/new"
+						component={RegLink}
+					>
 						NEW CLAIM
 					</Button>
 				</Grid>
@@ -28,5 +34,9 @@ const ClaimHeader = (props) => {
 		</React.Fragment>
 	);
 };
+
+const RegLink = React.forwardRef((props, ref) => (
+	<Link innerRef={ref} {...props} />
+));
 
 export default withStyles(styles, { withTheme: true })(ClaimHeader);
