+++ client/src/Dashboard/pages/Claims/components/Table/Table.js
@@ -1,30 +1,18 @@
 import React from "react";
-import clsx from "clsx";
-import PropTypes from "prop-types";
-import { lighten, makeStyles } from "@material-ui/core/styles";
+import { makeStyles } from "@material-ui/core/styles";
 import {
 	Table,
 	TableBody,
 	TableCell,
-	TableHead,
 	TablePagination,
 	TableRow,
-	TableSortLabel,
-	Toolbar,
-	Typography,
 	Paper,
-	Checkbox,
-	IconButton,
-	Tooltip,
-	FormControlLabel,
-	Switch
-} from "@material-ui/core/";
-import DeleteIcon from "@material-ui/icons/Delete";
-import FilterListIcon from "@material-ui/icons/FilterList";
+	Button
+} from "@material-ui/core";
 
-function createData(name, calories, fat, carbs, protein) {
-	return { name, calories, fat, carbs, protein };
-}
+import EnhancedTableToolbar from "./components/Toolbar";
+import EnhancedTableHead from "./components/TableHead";
+import { createData, getSorting, stableSort } from "./utils";
 
 const rows = [
 	createData("Cupcake", 305, 3.7, 67, 4.3),
@@ -41,170 +29,6 @@ const rows = [
 	createData("Nougat", 360, 19.0, 9, 37.0),
 	createData("Oreo", 437, 18.0, 63, 4.0)
 ];
-
-function desc(a, b, orderBy) {
-	if (b[orderBy] < a[orderBy]) {
-		return -1;
-	}
-	if (b[orderBy] > a[orderBy]) {
-		return 1;
-	}
-	return 0;
-}
-
-function stableSort(array, cmp) {
-	const stabilizedThis = array.map((el, index) => [el, index]);
-	stabilizedThis.sort((a, b) => {
-		const order = cmp(a[0], b[0]);
-		if (order !== 0) return order;
-		return a[1] - b[1];
-	});
-	return stabilizedThis.map((el) => el[0]);
-}
-
-function getSorting(order, orderBy) {
-	return order === "desc"
-		? (a, b) => desc(a, b, orderBy)
-		: (a, b) => -desc(a, b, orderBy);
-}
-
-const headRows = [
-	{
-		id: "name",
-		numeric: false,
-		disablePadding: true,
-		label: "Dessert (100g serving)"
-	},
-	{ id: "calories", numeric: true, disablePadding: false, label: "Calories" },
-	{ id: "fat", numeric: true, disablePadding: false, label: "Fat (g)" },
-	{ id: "carbs", numeric: true, disablePadding: false, label: "Carbs (g)" },
-	{ id: "protein", numeric: true, disablePadding: false, label: "Protein (g)" }
-];
-
-function EnhancedTableHead(props) {
-	const {
-		onSelectAllClick,
-		order,
-		orderBy,
-		numSelected,
-		rowCount,
-		onRequestSort
-	} = props;
-	const createSortHandler = (property) => (event) => {
-		onRequestSort(event, property);
-	};
-
-	return (
-		<TableHead>
-			<TableRow>
-				<TableCell padding="checkbox">
-					<Checkbox
-						indeterminate={numSelected > 0 && numSelected < rowCount}
-						checked={numSelected === rowCount}
-						onChange={onSelectAllClick}
-						inputProps={{ "aria-label": "Select all desserts" }}
-					/>
-				</TableCell>
-				{headRows.map((row) => (
-					<TableCell
-						key={row.id}
-						align={row.numeric ? "right" : "left"}
-						padding={row.disablePadding ? "none" : "default"}
-						sortDirection={orderBy === row.id ? order : false}
-					>
-						<TableSortLabel
-							active={orderBy === row.id}
-							direction={order}
-							onClick={createSortHandler(row.id)}
-						>
-							{row.label}
-						</TableSortLabel>
-					</TableCell>
-				))}
-			</TableRow>
-		</TableHead>
-	);
-}
-
-EnhancedTableHead.propTypes = {
-	numSelected: PropTypes.number.isRequired,
-	onRequestSort: PropTypes.func.isRequired,
-	onSelectAllClick: PropTypes.func.isRequired,
-	order: PropTypes.string.isRequired,
-	orderBy: PropTypes.string.isRequired,
-	rowCount: PropTypes.number.isRequired
-};
-
-const useToolbarStyles = makeStyles((theme) => ({
-	root: {
-		paddingLeft: theme.spacing(2),
-		paddingRight: theme.spacing(1)
-	},
-	highlight:
-		theme.palette.type === "light"
-			? {
-					color: theme.palette.secondary.main,
-					backgroundColor: lighten(theme.palette.secondary.light, 0.85)
-			  }
-			: {
-					color: theme.palette.text.primary,
-					backgroundColor: theme.palette.secondary.dark
-			  },
-	spacer: {
-		flex: "1 1 100%"
-	},
-	actions: {
-		color: theme.palette.text.secondary
-	},
-	title: {
-		flex: "0 0 auto"
-	}
-}));
-
-const EnhancedTableToolbar = (props) => {
-	const classes = useToolbarStyles();
-	const { numSelected } = props;
-
-	return (
-		<Toolbar
-			className={clsx(classes.root, {
-				[classes.highlight]: numSelected > 0
-			})}
-		>
-			<div className={classes.title}>
-				{numSelected > 0 ? (
-					<Typography color="inherit" variant="subtitle1">
-						{numSelected} selected
-					</Typography>
-				) : (
-					<Typography variant="h6" id="tableTitle">
-						Nutrition
-					</Typography>
-				)}
-			</div>
-			<div className={classes.spacer} />
-			<div className={classes.actions}>
-				{numSelected > 0 ? (
-					<Tooltip title="Delete">
-						<IconButton aria-label="Delete">
-							<DeleteIcon />
-						</IconButton>
-					</Tooltip>
-				) : (
-					<Tooltip title="Filter list">
-						<IconButton aria-label="Filter list">
-							<FilterListIcon />
-						</IconButton>
-					</Tooltip>
-				)}
-			</div>
-		</Toolbar>
-	);
-};
-
-EnhancedTableToolbar.propTypes = {
-	numSelected: PropTypes.number.isRequired
-};
 
 const useStyles = makeStyles((theme) => ({
 	root: {
@@ -227,44 +51,13 @@ export default function EnhancedTable() {
 	const classes = useStyles();
 	const [order, setOrder] = React.useState("asc");
 	const [orderBy, setOrderBy] = React.useState("calories");
-	const [selected, setSelected] = React.useState([]);
 	const [page, setPage] = React.useState(0);
-	const [dense, setDense] = React.useState(false);
 	const [rowsPerPage, setRowsPerPage] = React.useState(5);
 
 	function handleRequestSort(event, property) {
 		const isDesc = orderBy === property && order === "desc";
 		setOrder(isDesc ? "asc" : "desc");
 		setOrderBy(property);
-	}
-
-	function handleSelectAllClick(event) {
-		if (event.target.checked) {
-			const newSelecteds = rows.map((n) => n.name);
-			setSelected(newSelecteds);
-			return;
-		}
-		setSelected([]);
-	}
-
-	function handleClick(event, name) {
-		const selectedIndex = selected.indexOf(name);
-		let newSelected = [];
-
-		if (selectedIndex === -1) {
-			newSelected = newSelected.concat(selected, name);
-		} else if (selectedIndex === 0) {
-			newSelected = newSelected.concat(selected.slice(1));
-		} else if (selectedIndex === selected.length - 1) {
-			newSelected = newSelected.concat(selected.slice(0, -1));
-		} else if (selectedIndex > 0) {
-			newSelected = newSelected.concat(
-				selected.slice(0, selectedIndex),
-				selected.slice(selectedIndex + 1)
-			);
-		}
-
-		setSelected(newSelected);
 	}
 
 	function handleChangePage(event, newPage) {
@@ -275,26 +68,22 @@ export default function EnhancedTable() {
 		setRowsPerPage(+event.target.value);
 	}
 
-	const isSelected = (name) => selected.indexOf(name) !== -1;
-
 	const emptyRows =
 		rowsPerPage - Math.min(rowsPerPage, rows.length - page * rowsPerPage);
 
 	return (
 		<div className={classes.root}>
 			<Paper className={classes.paper}>
-				<EnhancedTableToolbar numSelected={selected.length} />
+				<EnhancedTableToolbar />
 				<div className={classes.tableWrapper}>
 					<Table
 						className={classes.table}
 						aria-labelledby="tableTitle"
-						size={dense ? "small" : "medium"}
+						size={"medium"}
 					>
 						<EnhancedTableHead
-							numSelected={selected.length}
 							order={order}
 							orderBy={orderBy}
-							onSelectAllClick={handleSelectAllClick}
 							onRequestSort={handleRequestSort}
 							rowCount={rows.length}
 						/>
@@ -302,25 +91,11 @@ export default function EnhancedTable() {
 							{stableSort(rows, getSorting(order, orderBy))
 								.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
 								.map((row, index) => {
-									const isItemSelected = isSelected(row.name);
 									const labelId = `enhanced-table-checkbox-${index}`;
 
 									return (
-										<TableRow
-											hover
-											onClick={(event) => handleClick(event, row.name)}
-											role="checkbox"
-											aria-checked={isItemSelected}
-											tabIndex={-1}
-											key={row.name}
-											selected={isItemSelected}
-										>
-											<TableCell padding="checkbox">
-												<Checkbox
-													checked={isItemSelected}
-													inputProps={{ "aria-labelledby": labelId }}
-												/>
-											</TableCell>
+										<TableRow hover tabIndex={-1} key={row.name}>
+											<TableCell padding="checkbox" />
 											<TableCell
 												component="th"
 												id={labelId}
@@ -329,10 +104,13 @@ export default function EnhancedTable() {
 											>
 												{row.name}
 											</TableCell>
-											<TableCell align="right">{row.calories}</TableCell>
-											<TableCell align="right">{row.fat}</TableCell>
-											<TableCell align="right">{row.carbs}</TableCell>
-											<TableCell align="right">{row.protein}</TableCell>
+											<TableCell align="left">{row.calories}</TableCell>
+											<TableCell align="left">{row.fat}</TableCell>
+											<TableCell align="left">{row.carbs}</TableCell>
+											<TableCell align="left">{row.protein}</TableCell>
+											<TableCell align="left">
+												<Button variant="contained">REVIEW</Button>
+											</TableCell>
 										</TableRow>
 									);
 								})}
