+++ client/src/Dashboard/pages/Claims/components/Table/components/Toolbar.js
@@ -1,0 +1,75 @@
+import React from "react";
+import clsx from "clsx";
+import { Typography, Toolbar, Tooltip, IconButton } from "@material-ui/core";
+import { lighten, makeStyles } from "@material-ui/core/styles";
+
+import DeleteIcon from "@material-ui/icons/Delete";
+import FilterListIcon from "@material-ui/icons/FilterList";
+const useToolbarStyles = makeStyles((theme) => ({
+	root: {
+		paddingLeft: theme.spacing(2),
+		paddingRight: theme.spacing(1)
+	},
+	highlight:
+		theme.palette.type === "light"
+			? {
+					color: theme.palette.secondary.main,
+					backgroundColor: lighten(theme.palette.secondary.light, 0.85)
+			  }
+			: {
+					color: theme.palette.text.primary,
+					backgroundColor: theme.palette.secondary.dark
+			  },
+	spacer: {
+		flex: "1 1 100%"
+	},
+	actions: {
+		color: theme.palette.text.secondary
+	},
+	title: {
+		flex: "0 0 auto"
+	}
+}));
+
+const EnhancedTableToolbar = (props) => {
+	const classes = useToolbarStyles();
+	const { numSelected } = props;
+
+	return (
+		<Toolbar
+			className={clsx(classes.root, {
+				[classes.highlight]: numSelected > 0
+			})}
+		>
+			<div className={classes.title}>
+				{numSelected > 0 ? (
+					<Typography color="inherit" variant="subtitle1">
+						{numSelected} selected
+					</Typography>
+				) : (
+					<Typography variant="h6" id="tableTitle">
+						All Claims
+					</Typography>
+				)}
+			</div>
+			<div className={classes.spacer} />
+			<div className={classes.actions}>
+				{numSelected > 0 ? (
+					<Tooltip title="Delete">
+						<IconButton aria-label="Delete">
+							<DeleteIcon />
+						</IconButton>
+					</Tooltip>
+				) : (
+					<Tooltip title="Filter list">
+						<IconButton aria-label="Filter list">
+							<FilterListIcon />
+						</IconButton>
+					</Tooltip>
+				)}
+			</div>
+		</Toolbar>
+	);
+};
+
+export default EnhancedTableToolbar;
