+++ client/src/Dashboard/pages/Claims/pages/ClaimNew.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const ClaimNew = () => {
+	return (
+		<div>
+			<h1>Claim New</h1>
+		</div>
+	);
+};
+
+export default ClaimNew;
