+++ client/src/Dashboard/pages/Claims/pages/ClaimReview.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const ClaimReview = () => {
+	return (
+		<div>
+			<h1>Claim Review</h1>
+		</div>
+	);
+};
+
+export default ClaimReview;
